Consolidate project into single script and paper
