--- paper/REVISION_NOTES.md
@@ -1,9 +1,0 @@
-# Revision Notes — 2025-08-18
-
-- Introduced a **runtime-first** architecture with 12 enforceable primitives (CPCS, LoT, RoI‑S, IGMS, TTC, PSC‑CR, KVT‑P, IDS, CSS, SCCH, POPB, CS‑WIF).
-- Added concrete **interfaces, enforcement hooks, and acceptance tests** for CPCS, RoI‑S, and TTC.
-- Included **code snippets** for RoI Splicing and Two‑Phase Tool Calls.
-- Added a **compute-budgeted Growing RBF network** experiment and snippet demonstrating runtime growth gating.
-- Outlined a **fast validation plan** with measurable metrics.
-- Clarified why a “smart LLM allocator” cannot subsume these mechanisms: they require **OS/runtime/hardware** control and **verifier‑bound signals**.
-- Added a concise **abstract** highlighting the twelve runtime‑enforced primitives.
