--- paper/REVISION_NOTES.tex
@@ -1,88 +1,0 @@
-% Options for packages loaded elsewhere
-\PassOptionsToPackage{unicode}{hyperref}
-\PassOptionsToPackage{hyphens}{url}
-%
-\documentclass[
-]{article}
-\usepackage{amsmath,amssymb}
-\usepackage{iftex}
-\ifPDFTeX
-  \usepackage[T1]{fontenc}
-  \usepackage[utf8]{inputenc}
-  \usepackage{textcomp} % provide euro and other symbols
-\else % if luatex or xetex
-  \usepackage{unicode-math} % this also loads fontspec
-  \defaultfontfeatures{Scale=MatchLowercase}
-  \defaultfontfeatures[\rmfamily]{Ligatures=TeX,Scale=1}
-\fi
-\usepackage{lmodern}
-\ifPDFTeX\else
-  % xetex/luatex font selection
-\fi
-% Use upquote if available, for straight quotes in verbatim environments
-\IfFileExists{upquote.sty}{\usepackage{upquote}}{}
-\IfFileExists{microtype.sty}{% use microtype if available
-  \usepackage[]{microtype}
-  \UseMicrotypeSet[protrusion]{basicmath} % disable protrusion for tt fonts
-}{}
-\makeatletter
-\@ifundefined{KOMAClassName}{% if non-KOMA class
-  \IfFileExists{parskip.sty}{%
-    \usepackage{parskip}
-  }{% else
-    \setlength{\parindent}{0pt}
-    \setlength{\parskip}{6pt plus 2pt minus 1pt}}
-}{% if KOMA class
-  \KOMAoptions{parskip=half}}
-\makeatother
-\usepackage{xcolor}
-\setlength{\emergencystretch}{3em} % prevent overfull lines
-\providecommand{\tightlist}{%
-  \setlength{\itemsep}{0pt}\setlength{\parskip}{0pt}}
-\setcounter{secnumdepth}{-\maxdimen} % remove section numbering
-\ifLuaTeX
-  \usepackage{selnolig}  % disable illegal ligatures
-\fi
-\IfFileExists{bookmark.sty}{\usepackage{bookmark}}{\usepackage{hyperref}}
-\IfFileExists{xurl.sty}{\usepackage{xurl}}{} % add URL line breaks if available
-\urlstyle{same}
-\hypersetup{
-  hidelinks,
-  pdfcreator={LaTeX via pandoc}}
-
-\author{}
-\date{}
-
-\begin{document}
-
-\hypertarget{revision-notes-2025-08-18}{%
-\section{Revision Notes ---
-2025-08-18}\label{revision-notes-2025-08-18}}
-
-\begin{itemize}
-\tightlist
-\item
-  Introduced a \textbf{runtime-first} architecture with 12 enforceable
-  primitives (CPCS, LoT, RoI‑S, IGMS, TTC, PSC‑CR, KVT‑P, IDS, CSS,
-  SCCH, POPB, CS‑WIF).
-\item
-  Added concrete \textbf{interfaces, enforcement hooks, and acceptance
-  tests} for CPCS, RoI‑S, and TTC.
-\item
-  Included \textbf{code snippets} for RoI Splicing and Two‑Phase Tool
-  Calls.
-\item
-  Added a \textbf{compute-budgeted Growing RBF network} experiment and
-  snippet demonstrating runtime growth gating.
-\item
-  Outlined a \textbf{fast validation plan} with measurable metrics.
-\item
-  Clarified why a ``smart LLM allocator'' cannot subsume these
-  mechanisms: they require \textbf{OS/runtime/hardware} control and
-  \textbf{verifier‑bound signals}.
-\item
-  Added a concise \textbf{abstract} highlighting the twelve
-  runtime‑enforced primitives.
-\end{itemize}
-
-\end{document}
