--- paper/specs/RoIS_spec.tex
@@ -1,107 +1,0 @@
-% Options for packages loaded elsewhere
-\PassOptionsToPackage{unicode}{hyperref}
-\PassOptionsToPackage{hyphens}{url}
-%
-\documentclass[
-]{article}
-\usepackage{amsmath,amssymb}
-\usepackage{iftex}
-\ifPDFTeX
-  \usepackage[T1]{fontenc}
-  \usepackage[utf8]{inputenc}
-  \usepackage{textcomp} % provide euro and other symbols
-\else % if luatex or xetex
-  \usepackage{unicode-math} % this also loads fontspec
-  \defaultfontfeatures{Scale=MatchLowercase}
-  \defaultfontfeatures[\rmfamily]{Ligatures=TeX,Scale=1}
-\fi
-\usepackage{lmodern}
-\ifPDFTeX\else
-  % xetex/luatex font selection
-\fi
-% Use upquote if available, for straight quotes in verbatim environments
-\IfFileExists{upquote.sty}{\usepackage{upquote}}{}
-\IfFileExists{microtype.sty}{% use microtype if available
-  \usepackage[]{microtype}
-  \UseMicrotypeSet[protrusion]{basicmath} % disable protrusion for tt fonts
-}{}
-\makeatletter
-\@ifundefined{KOMAClassName}{% if non-KOMA class
-  \IfFileExists{parskip.sty}{%
-    \usepackage{parskip}
-  }{% else
-    \setlength{\parindent}{0pt}
-    \setlength{\parskip}{6pt plus 2pt minus 1pt}}
-}{% if KOMA class
-  \KOMAoptions{parskip=half}}
-\makeatother
-\usepackage{xcolor}
-\setlength{\emergencystretch}{3em} % prevent overfull lines
-\providecommand{\tightlist}{%
-  \setlength{\itemsep}{0pt}\setlength{\parskip}{0pt}}
-\setcounter{secnumdepth}{-\maxdimen} % remove section numbering
-\ifLuaTeX
-  \usepackage{selnolig}  % disable illegal ligatures
-\fi
-\IfFileExists{bookmark.sty}{\usepackage{bookmark}}{\usepackage{hyperref}}
-\IfFileExists{xurl.sty}{\usepackage{xurl}}{} % add URL line breaks if available
-\urlstyle{same}
-\hypersetup{
-  hidelinks,
-  pdfcreator={LaTeX via pandoc}}
-
-\author{}
-\date{}
-
-\begin{document}
-
-\hypertarget{roi-s-region-of-interest-splicing-spec}{%
-\section{RoI-S --- Region-of-Interest Splicing
-(Spec)}\label{roi-s-region-of-interest-splicing-spec}}
-
-\textbf{Goal.} Repair only the erroneous span without losing
-prefix/suffix caches.
-
-\hypertarget{interfaces}{%
-\subsection{Interfaces}\label{interfaces}}
-
-\begin{itemize}
-\tightlist
-\item
-  \texttt{mark\_span(doc\_id,\ start\_token,\ end\_token)} from verifier
-\item
-  \texttt{partial\_decode(model\_id,\ kv\_ckpt,\ prefix\_tokens,\ target\_end)}
-  returns \texttt{repair\_tokens}
-\item
-  \texttt{splice(doc\_id,\ start,\ end,\ repair\_tokens)} updates
-  document and KV-store
-\end{itemize}
-
-\hypertarget{enforcement}{%
-\subsection{Enforcement}\label{enforcement}}
-
-\begin{itemize}
-\tightlist
-\item
-  Decoder exposes KV slice/extend primitives; deterministic stitch
-  verified by hash of (prefix, repair, suffix)
-\item
-  Abort if splice invalidates verifier hash
-\end{itemize}
-
-\hypertarget{acceptance-tests}{%
-\subsection{Acceptance Tests}\label{acceptance-tests}}
-
-\begin{enumerate}
-\def\labelenumi{\arabic{enumi}.}
-\tightlist
-\item
-  \textbf{Token Savings:} ≥60\% fewer regenerated tokens per fix vs full
-  re-decode
-\item
-  \textbf{Determinism:} identical outputs across 5 runs with fixed RNG
-\item
-  \textbf{Latency:} ≤40\% of full re-decode wall time for 256-token RoIs
-\end{enumerate}
-
-\end{document}
