--- paper/specs/TTC_spec.tex
@@ -1,109 +1,0 @@
-% Options for packages loaded elsewhere
-\PassOptionsToPackage{unicode}{hyperref}
-\PassOptionsToPackage{hyphens}{url}
-%
-\documentclass[
-]{article}
-\usepackage{amsmath,amssymb}
-\usepackage{iftex}
-\ifPDFTeX
-  \usepackage[T1]{fontenc}
-  \usepackage[utf8]{inputenc}
-  \usepackage{textcomp} % provide euro and other symbols
-\else % if luatex or xetex
-  \usepackage{unicode-math} % this also loads fontspec
-  \defaultfontfeatures{Scale=MatchLowercase}
-  \defaultfontfeatures[\rmfamily]{Ligatures=TeX,Scale=1}
-\fi
-\usepackage{lmodern}
-\ifPDFTeX\else
-  % xetex/luatex font selection
-\fi
-% Use upquote if available, for straight quotes in verbatim environments
-\IfFileExists{upquote.sty}{\usepackage{upquote}}{}
-\IfFileExists{microtype.sty}{% use microtype if available
-  \usepackage[]{microtype}
-  \UseMicrotypeSet[protrusion]{basicmath} % disable protrusion for tt fonts
-}{}
-\makeatletter
-\@ifundefined{KOMAClassName}{% if non-KOMA class
-  \IfFileExists{parskip.sty}{%
-    \usepackage{parskip}
-  }{% else
-    \setlength{\parindent}{0pt}
-    \setlength{\parskip}{6pt plus 2pt minus 1pt}}
-}{% if KOMA class
-  \KOMAoptions{parskip=half}}
-\makeatother
-\usepackage{xcolor}
-\setlength{\emergencystretch}{3em} % prevent overfull lines
-\providecommand{\tightlist}{%
-  \setlength{\itemsep}{0pt}\setlength{\parskip}{0pt}}
-\setcounter{secnumdepth}{-\maxdimen} % remove section numbering
-\ifLuaTeX
-  \usepackage{selnolig}  % disable illegal ligatures
-\fi
-\IfFileExists{bookmark.sty}{\usepackage{bookmark}}{\usepackage{hyperref}}
-\IfFileExists{xurl.sty}{\usepackage{xurl}}{} % add URL line breaks if available
-\urlstyle{same}
-\hypersetup{
-  hidelinks,
-  pdfcreator={LaTeX via pandoc}}
-
-\author{}
-\date{}
-
-\begin{document}
-
-\hypertarget{ttc-transactional-tool-calls-spec}{%
-\section{TTC --- Transactional Tool Calls
-(Spec)}\label{ttc-transactional-tool-calls-spec}}
-
-\textbf{Goal.} Prevent orphaned heavy runs and sunk compute when plans
-change.
-
-\hypertarget{interfaces}{%
-\subsection{Interfaces}\label{interfaces}}
-
-\begin{itemize}
-\tightlist
-\item
-  \texttt{prepare(inputs,\ est\_cost,\ consumers{[}{]})\ -\textgreater{}\ call\_id}
-\item
-  \texttt{will\_consume(call\_id)\ -\textgreater{}\ bool} (coordinator
-  arbitration)
-\item
-  \texttt{commit(call\_id)\ -\textgreater{}\ result}
-\item
-  \texttt{abort(call\_id)}
-\end{itemize}
-
-\hypertarget{enforcement}{%
-\subsection{Enforcement}\label{enforcement}}
-
-\begin{itemize}
-\tightlist
-\item
-  Tools must block heavy execution until \texttt{commit}
-\item
-  Coordinator maintains a durable log: \texttt{PREPARE},
-  \texttt{COMMIT}, \texttt{ABORT}
-\end{itemize}
-
-\hypertarget{acceptance-tests}{%
-\subsection{Acceptance Tests}\label{acceptance-tests}}
-
-\begin{enumerate}
-\def\labelenumi{\arabic{enumi}.}
-\tightlist
-\item
-  \textbf{Orphan Rate:} \textasciitilde0 across forced plan changes
-  (≥100 randomized trials)
-\item
-  \textbf{Bounded Overhead:} \texttt{prepare} adds ≤2\% latency vs
-  direct call
-\item
-  \textbf{Auditability:} replay log reconstructs decisions exactly
-\end{enumerate}
-
-\end{document}
