--- results/bandit_plusmlp_summary.csv
@@ -1,4 +1,0 @@
-Policy,Final Cumulative Reward,Final Cumulative Regret,Final #Prototypes,Growth Budget Remaining
-GRBFN+Plastic,6190.0000,239.1557,22.0000,0.0000
-LinUCB,6152.0000,299.4125,,
-MLPOnline,6291.0000,247.6138,,
