--- results/bandit_summary.csv
@@ -1,3 +1,0 @@
-Policy,Final Cumulative Reward,Final Cumulative Regret,Final #Prototypes,Growth Budget Remaining
-GRBFN+Plastic,6194.0000,262.0921,23.0000,0.0000
-LinUCB,6256.0000,222.6492,,
