--- results/bandit_summary_plusmlp.csv
@@ -1,4 +1,0 @@
-Policy,Final Cumulative Reward,Final Cumulative Regret,Final #Prototypes (GRBFN+)
-LinUCB(alpha=0.8),6152.0,299.41248174071654,
-GRBFN+Plastic,6190.0,239.15574626621986,22.0
-MLP(32),6291.0,247.61384443497002,
