--- results/classification_postdrift.csv
@@ -1,4 +1,0 @@
-Model,Post@200(avg),Post@500(avg)
-GRBFN+Plastic,0.815,0.8716
-LogisticSGD,0.65,0.7144
-MLPOnline,0.637,0.7344
